requirements.txt
#mcp[cli]==1.8.1  #manually copied into app package (with modifications)
semantic-kernel==1.31.0

azure-keyvault-secrets
azure-cosmos
azure-identity
azure-mgmt-web
azure-mgmt-storage
azure-storage-blob
azure.appconfiguration
azure.appconfiguration.provider
azure-monitor-opentelemetry
azure-monitor-opentelemetry-exporter
azure-cognitiveservices-speech

tenacity
uv

fastapi
#fastapi-mcp==0.3.4
sse_starlette
httpx_sse

gunicorn
uvicorn

python-dotenv

#tools support
autogen-agentchat==0.5.7
autogen_ext==0.5.7
tiktoken
sqlparse
azure-servicebus
azure-eventgrid
#git+https://github.com/givenscj/graphrag.git#egg=graphrag
#graphrag

#excel
openpyxl

#word
python-docx

#databricks
pyspark==3.4.4
#databricks-sdk
#databricks-connect==12.2.45
#databricks-sql-connector
pandas

#code interpreter
azure-ai-projects

#presidio
presidio_analyzer 
presidio_anonymizer

##wikipedia
wikipedia

#postgresql
psycopg[binary]
pyodbc

##Bing
azure-cognitiveservices-search-websearch

#neo4j
neo4j-graphrag
neo4j-graphrag[openai]